+++ R/RcppExports.R
@@ -117,6 +117,10 @@
     .Call(`_Cyclops_cyclopsRunBootstrap`, inRcppCcdInterface, outFileName, treatmentId, replicates)
 }
 
+.cyclopsGetLogLikelihoodGradient <- function(inRcppCcdInterface) {
+    .Call(`_Cyclops_cyclopsGetLogLikelihoodGradient`, inRcppCcdInterface)
+}
+
 .cyclopsLogModel <- function(inRcppCcdInterface) {
     .Call(`_Cyclops_cyclopsLogModel`, inRcppCcdInterface)
 }