+++ NEWS.md
@@ -4,6 +4,7 @@ develop
 Changes:
 
 1. fix some NaN/-Inf inconsistencies with likelihood profiling
+2. make `maxRety` a function parameter
 
 Cyclops v3.4.0
 ==============